requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9
celery[redis]>=5.3
redis>=5.0
openai>=1.30
watchdog>=4.0
python-dotenv>=1.0
gunicorn>=21.0
django-celery-results>=2.5
djangorestframework>=3.15
django-cors-headers>=4.3
